Added heartbeat support for rest endpoints. Check notifications count on heartbeat.
